Move new task page + fix colors
